[feat][#36] 홈 뷰모델 분리
